# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  utils.SuccessResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@shiftley.com
    name: API Support
    url: http://www.shiftley.com/support
  description: Shiftley Backend API Documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Shiftley API
  version: "1.0"
paths:
...
  /verifier/employees/{id}/verify:
    post:
      consumes:
      - application/json
      description: Remotely approve an employee after reviewing Aadhaar and photo.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Status and Notes
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Verify Employee (Remote KYC)
      tags:
      - Verifier
  /verifier/employers/{id}/verify:
    post:
      consumes:
      - multipart/form-data
      description: Finalizes the physical location check with photographic proof.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: APPROVED or REJECTED
        in: formData
        name: status
        required: true
        type: string
      - description: Verification Notes
        in: formData
        name: notes
        type: string
      - description: JSON GPS pin override
        in: formData
        name: verified_location
        type: string
      - description: Selfie of verifier at location
        in: formData
        name: verifier_selfie
        required: true
        type: file
      - description: Location Proof 1
        in: formData
        name: location_photo_1
        required: true
        type: file
      - description: Location Proof 2
        in: formData
        name: location_photo_2
        required: true
        type: file
      - description: Location Proof 3
        in: formData
        name: location_photo_3
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Verify Employer (Physical Visit)
      tags:
      - Verifier
  /verifier/queue:
    get:
      description: Fetches a list of users awaiting verification.
      parameters:
      - description: User Type (EMPLOYER or EMPLOYEE)
        in: query
        name: type
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Pending Queues
      tags:
      - Verifier
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"